* base Brainrot Evo

+++ Assets/Scripts/Event/Event_Player_Add_Exp.cs
@@ -5,9 +5,9 @@ namespace Game
 {
     public class Event_Player_Add_Exp : IEvent
     {
-        public float exp;
+        public int exp;
 
-        public Event_Player_Add_Exp(float exp)
+        public Event_Player_Add_Exp(int exp)
         {
             this.exp = exp;
         }
